+++ ui/apps/platform/src/Containers/Clusters/Components/CollectorStatus.js
@@ -3,14 +3,11 @@ import React from 'react';
 
 import DetailedTooltipOverlay from 'Components/DetailedTooltipOverlay';
 import Tooltip from 'Components/Tooltip';
-import TooltipOverlay from 'Components/TooltipOverlay';
 import { healthStatusLabels } from 'messages/common';
-import { getDistanceStrict } from 'utils/dateUtils';
+import { getDistanceStrictAsPhrase } from 'utils/dateUtils';
 
 import HealthStatus from './HealthStatus';
 import { healthStatusStyles, isDelayedSensorHealthStatus } from '../cluster.helpers';
-
-const reasonUnavailable = 'Sensor version is too out of date';
 
 const trClassName = 'align-bottom leading-normal'; // align-bottom in case heading text wraps
 const thClassName = 'font-600 pl-0 pr-1 py-0 text-left';
@@ -24,9 +21,10 @@ const tdClassName = 'p-0 text-right';
 const CollectorStatus = ({
     collectorHealthStatus,
     collectorHealthInfo,
+    healthInfoComplete,
     sensorHealthStatus,
     lastContact,
-    now,
+    currentDatetime,
     isList,
 }) => {
     if (collectorHealthStatus) {
@@ -37,93 +35,88 @@ const CollectorStatus = ({
             </span>
         );
 
-        if (collectorHealthStatus === 'UNAVAILABLE') {
-            return isList ? (
-                <Tooltip content={<TooltipOverlay>{reasonUnavailable}</TooltipOverlay>}>
-                    <div>
-                        <HealthStatus Icon={Icon} iconColor={fgColor}>
-                            {labelElement}
-                        </HealthStatus>
-                    </div>
-                </Tooltip>
-            ) : (
-                <HealthStatus Icon={Icon} iconColor={fgColor}>
-                    <div>
-                        {labelElement}
-                        <br />
-                        {reasonUnavailable}
-                    </div>
-                </HealthStatus>
-            );
-        }
-
         // In rare case that the block does not fit in a narrow column,
         // the space and "whitespace-no-wrap" cause time phrase to wrap as a unit.
+        // Order arguments according to date-fns@2 convention:
+        // If lastContact <= currentDateTime: X units ago
         const statusElement =
             lastContact && isDelayedSensorHealthStatus(sensorHealthStatus) ? (
                 <div>
                     {labelElement}{' '}
-                    <span className="whitespace-no-wrap">{`${getDistanceStrict(
-                        lastContact,
-                        now
-                    )} ago`}</span>
+                    <span className="whitespace-no-wrap">
+                        {getDistanceStrictAsPhrase(lastContact, currentDatetime)}
+                    </span>
                 </div>
             ) : (
                 <div>{labelElement}</div>
             );
 
-        if (collectorHealthInfo) {
-            const { totalReadyPods, totalDesiredPods, totalRegisteredNodes } = collectorHealthInfo;
-            const totalsElement = (
-                <table>
-                    <tbody>
-                        <tr className={trClassName} key="totalReadyPods">
-                            <th className={thClassName} scope="row">
-                                Collector pods ready:
-                            </th>
-                            <td className={tdClassName}>{totalReadyPods}</td>
-                        </tr>
-                        <tr className={trClassName} key="totalDesiredPods">
-                            <th className={thClassName} scope="row">
-                                Collector pods expected:
-                            </th>
-                            <td className={tdClassName}>{totalDesiredPods}</td>
-                        </tr>
-                        <tr className={trClassName} key="totalRegisteredNodes">
-                            <th className={thClassName} scope="row">
-                                Registered nodes in cluster:
-                            </th>
-                            <td className={tdClassName}>{totalRegisteredNodes}</td>
-                        </tr>
-                    </tbody>
-                </table>
-            );
+        const healthInfoElement = collectorHealthInfo ? (
+            <table>
+                <tbody>
+                    <tr className={trClassName} key="totalReadyPods">
+                        <th className={thClassName} scope="row">
+                            Collector pods ready:
+                        </th>
+                        <td className={tdClassName}>
+                            <span className={`${bgColor} ${fgColor}`}>
+                                {collectorHealthInfo.totalReadyPods}
+                            </span>
+                        </td>
+                    </tr>
+                    <tr className={trClassName} key="totalDesiredPods">
+                        <th className={thClassName} scope="row">
+                            Collector pods expected:
+                        </th>
+                        <td className={tdClassName}>
+                            <span className={`${bgColor} ${fgColor}`}>
+                                {collectorHealthInfo.totalDesiredPods}
+                            </span>
+                        </td>
+                    </tr>
+                    <tr className={trClassName} key="totalRegisteredNodes">
+                        <th className={thClassName} scope="row">
+                            Registered nodes in cluster:
+                        </th>
+                        <td className={tdClassName}>{collectorHealthInfo.totalRegisteredNodes}</td>
+                    </tr>
+                </tbody>
+            </table>
+        ) : null;
 
-            return isList ? (
-                <Tooltip
-                    content={
-                        <DetailedTooltipOverlay title="Collector DaemonSet" body={totalsElement} />
-                    }
-                >
-                    <div>
-                        <HealthStatus Icon={Icon} iconColor={fgColor}>
-                            {statusElement}
-                        </HealthStatus>
-                    </div>
-                </Tooltip>
+        const incompletenessElement =
+            collectorHealthStatus !== 'UNINITIALIZED' && healthInfoComplete === false ? (
+                <strong>Upgrade sensor to get complete collector health information</strong>
+            ) : null;
+
+        const auxiliaryElement =
+            healthInfoElement && incompletenessElement ? (
+                <div>
+                    {healthInfoElement}
+                    {incompletenessElement}
+                </div>
             ) : (
-                <HealthStatus Icon={Icon} iconColor={fgColor}>
-                    <div>
-                        {statusElement}
-                        {totalsElement}
-                    </div>
-                </HealthStatus>
+                healthInfoElement || incompletenessElement
             );
-        }
 
-        return (
+        return isList && auxiliaryElement ? (
+            <Tooltip
+                content={
+                    <DetailedTooltipOverlay title="Collector DaemonSet" body={auxiliaryElement} />
+                }
+            >
+                <div>
+                    <HealthStatus Icon={Icon} iconColor={fgColor}>
+                        {statusElement}
+                    </HealthStatus>
+                </div>
+            </Tooltip>
+        ) : (
             <HealthStatus Icon={Icon} iconColor={fgColor}>
-                {statusElement}
+                <div>
+                    {statusElement}
+                    {auxiliaryElement}
+                </div>
             </HealthStatus>
         );
     }
@@ -144,15 +137,17 @@ CollectorStatus.propTypes = {
         totalReadyPods: PropTypes.number.isRequired,
         totalRegisteredNodes: PropTypes.number.isRequired,
     }),
+    healthInfoComplete: PropTypes.bool,
     sensorHealthStatus: PropTypes.oneOf(['UNINITIALIZED', 'UNHEALTHY', 'DEGRADED', 'HEALTHY']),
     lastContact: PropTypes.string, // ISO 8601
-    now: PropTypes.instanceOf(Date).isRequired,
+    currentDatetime: PropTypes.instanceOf(Date).isRequired,
     isList: PropTypes.bool.isRequired,
 };
 
 CollectorStatus.defaultProps = {
     collectorHealthStatus: null,
     collectorHealthInfo: null,
+    healthInfoComplete: false,
     sensorHealthStatus: null,
     lastContact: null,
 };